Refactor to make the code a bit nicer
